+++ CHANGELOG.md
@@ -2,12 +2,21 @@
 
 Combined changelog for GoLogin node.js SDK
 
-## [2.1.23] 2025-06-02
+## [2.1.23] 2025-06-09
 
+
+### Fixes
+
+* Viewport of puppeteer complies with profile resolution
+
+### New
+
+* Added fallback url for blocked countries
 
 ### Miscellaneous Chores
 
 * Added sentry to have information about errors that users face
+
 
 ## [2.1.21] 2025-04-23